Refactoring: move setLanguage(),__() to I18n.php
