+++ app/core/common_functions.php
@@ -13,7 +13,6 @@ function h($text)
   return htmlentities($text, ENT_QUOTES, \Fc2blog\Config::get('INTERNAL_ENCODING'));
 }
 
-
 /**
  * マルチバイト対応のtruncate
  */
@@ -63,15 +62,6 @@ function df($date, $format = 'Y/m/d H:i:s')
   return date($format, strtotime($date));
 }
 
-
-/**
- * 多言語化 TODO:後で独自実装関数とすげ替える予定
- */
-function __($msg)
-{
-  return _($msg);
-}
-
 /**
  * パスカルケースへの変換
  */
@@ -97,32 +87,6 @@ function camelCase($snake_case)
 function snakeCase($camel_case)
 {
   return strtolower(preg_replace("/([A-Z])/u", "_$0", lcfirst($camel_case)));
-}
-
-/**
- * 言語設定
- * @param \Fc2blog\Web\Request $request
- * @return string
- */
-function setLanguage(\Fc2blog\Web\Request $request)
-{
-  $cookie_lang = \Fc2blog\Web\Cookie::get('lang');
-  $cookie_language = \Fc2blog\Config::get('LANGUAGES.' . (string)$cookie_lang);
-  if (!is_null($cookie_lang) && !is_null($cookie_language)) {
-    $lang = $cookie_lang;
-    $language = $cookie_language;
-  }else{
-    $lang = \Fc2blog\Config::get('LANG');
-    $language = \Fc2blog\Config::get('LANGUAGE');
-  }
-
-  // 多言語化対応
-  putenv('LANG=' . $language);
-  putenv('LANGUAGE=' . $language);
-  setlocale(LC_ALL, $language);
-  bindtextdomain('messages', \Fc2blog\Config::get('LOCALE_DIR'));
-  textdomain('messages');
-  return $lang;
 }
 
 function getServerUrl()
